Retrieval of type information for actual values
There's a difference between "a type that can be stored in a data node"
(which is available through the associated schema node) and the "type
that represents the current value of this terminal data node". The
latter type information was not available through the C++ bindings, so
this patch makes that thing accessible.

Unfortunately, libyang does not use the YANG schema directly; instead,
it works with its "compiled representation" which hides some info, such
as the actual type names. This is compensated for, to a certain extent,
through the PrivParsed context option, but that feature does not help in
this particular case.

As a result, while one can retrieve the type information of the actual
data value, the set of information that is available is very, very
limited. It seems that the "plugin ID" (which is something that's only
used internally in libyang) is the only usable info.

The original motivation behind this feature of the C++ bindings was
distinguishing between IPv6 address literals and "anything else" in
hostnames, as implemented by the ietf-inet-types YANG module, and it
looks like this hack is "enough". I'm not happy with "leaking" such an
internal information from libayng through the C++ API, but I guess it's
better than parsing the string values again and guessing whether they
are IPv6 addresses.

Change-Id: Ic10d1bae6488055df1c9660c23102eee4ea2e3d8