Accessing global module typedefs
A libyang::Typedef is different from a libayng::types::Type because the
typedef itself has a name, description, units, etc, which might be
useful. So, a typedef is a construct from the raw, uncompiled YANG
model, whereas the types::Type is something which imposes some
constraints on data values.

This patch only adds support for named, module-global typedefs. Since
typedefs can be potentially defined below almost any node, those cannot
be retrieved via this function.

Also, some of the types retrieved through this won't have the "compiled"
`lysc_type`. This is OK-ish for builtin types, I suppose, but in the
real world I'm also getting this for unions:

Change-Id: I8d4093653358f17de71f8a6af819896c3c8423e5