version bump for libyang v2.2
First of all, it used to be called v3, but it got released as v2.2, so
let's make that clear. Also, since we've cleaned up our API a bit, let's
mark this one as v2. Hopefully this should make it clear that we do not
implement any semver guarantees and that we are not committing to stable
API/ABI just yet.

Change-Id: I6a0c5deb605597ef8d27120be9376761b9b2e904