Check if node with default value was created implicitly
In the RESTCONF server we should detect if the leaf which has a default
value was created implicitly or explicitly. The RFC [1, sec. 3.5.4]
states:

  If the target of a GET method is a data node that represents a leaf
  or leaf-list that has a default value and the leaf or leaf-list has
  not been instantiated yet (...)

libyang's lyd_is_default which is wrapped with
DataNodeTerm::hasDefaultValue only compares by value and does not care
about how was the value created. Fortunately, the implicitly created
nodes have a flag LYD_DEFAULT set [2].

  The implicit (default) nodes, created by libyang, are marked with
  the LYD_DEFAULT flag in lyd_node.flags member

[1] https://www.rfc-editor.org/rfc/rfc8040.html#section-3.5.4
[2] https://github.com/CESNET/libyang/blob/f8c0b58844aa9157ea90091a4a0f24515399627f/src/tree_data.h#L386

Change-Id: I362060b77639e88ff90c26b8467e469e87ef77f7