Rename DataNodeTerm::isDefaultValue to hasDefaultValue
This method wraps lyd_is_default which only compares the values. It does
not return any info about explicit/implicit creation of the node.

In the next commits we will have to differentiate between explicitly and
implicitly created nodes with default value. This change will make the
interface more clear.

Change-Id: Ibf93d8b427fcfd84bb44a38bd338ce0bba9a5b84