compare Module instances
I want to compare if a meta entry's module is the same as a module I
have.

The comparison is implemented on the value of lys_module*. It seems that
successive calls to ly_ctx_get_module always return the same pointer so
it *should* be safe. However, same modules in different contexts compare
as not equal.

Change-Id: I96efca783d5e780e98693330803e73847151d7af