Assign collection to Collection::end() iterator
In the current code, whenever I try to compare two Collection::end() I
receive an exception saying "Iterator is invalid". This is caused by the
fact that that operator== checks whether the current iterator has a
collection assigned.
However, the end() iterator was always constructed without the pointer
to the collection.

The fix assigns collection to the end() iterator. Because end() iterator
is represented as a nullptr, two end() iterators to two different
collections would compare as identical. Therefore, the Iterator::op==
now also takes the collection into account.

Change-Id: I77dc05a11a2ce610317ab14615cc619b07dd1a07