Allow parsing standalone notifications in Context::parseOp
Even when the libyang docs is not explicit on that, lyd_parse_op is used
for parsing standalone notifications too.

For such case, the lyd_node* tree and op are the same, or, they point
into the nodes in the same tree. So we should not wrap the pointers
separately (this would result in a heap-after-use).

Change-Id: Ib94114193a1a9073d9a4c6d15522eb41ba9c425d