add utility function for formatting time_points in yang:date-and-time
Many of our projects implement conversion of a time_point into
a yang:date-and-time string. It makes no sense to have this code
scattered across our codebase, let us implement in in one place, in
libyang-cpp.

We use 3rd party lib (HowardHinnant/date [1]) for the serialization of
timezone aware dates. Since this new utility header will probably
not be used frequently, we do not require it to be present when
installing libyang-cpp.
Instead, we provide it in a new pkgconfig entry.

[1] https://github.com/HowardHinnant/date

Change-Id: I51242fe2ec873fba5954c44e38e898d205a9e484