add utility function for formatting time_points in yang:date-and-time
Many of our projects implement conversion of a time_point into
a yang:date-and-time string. It makes no sense to have this code
scattered across our codebase, let us implement in in one place, in
libyang-cpp.

This is a header only implementation.

The problem is that std::chrono stream implementations are C++20 feature
but they might not implemented in the compiler (e.g., gcc implements
this in version 13 and 14). The header tries to detect if your compiler
supports the needed std::chrono features and in case it does not it
tries to include headers from the external HowardHinnant/date library[1]
(you have to link your executable to the library yourself, there is no
configuration for that).

[1] https://github.com/HowardHinnant/date

Change-Id: I51242fe2ec873fba5954c44e38e898d205a9e484