Add `Module::child()`, `Module::childrenDfs()` and `Module::immediateChildren()`
This function are implemented in the same manner as in `SchemaNode` and allows to walk through modules children. This is counterpart to already implemented `Module::childInstantiables()` that returns instantibles schema nodes. Those returns origin nodes, so nodes like choice and case can be accessed if end-user needs to read its schema.

While the implementation is inspired by functions in `SchemaNode`, imlementation of `Module::parent()` and `Module::siblings()` was omitted as those do no make sense on `Module`.

Change-Id: I38c8374304f859d65343d04d08302e07deb05f27