Extend implementation of `SchemaNode::child()` and improve its and `SchemaNode::childInstantiables` tests
Until now, `SchemaNode::child()` was returning all nodes except
`action` and `notification` because of lack of support. I implemented it
 in the same manner as it is done in libyang. Unfortunately, there
is no libyang function for accessing single child that would return
also actions or notifications if other nodes are not present and in
correct order. There is only function for DFS traversal.

This is preparation for extending `SchemaNode::siblings` that internally
uses `SchemaNode::child()` and due to which
`SchemaNode::immediateChildren` and `SchemaNode::childrenDfs`
do not return `action` and `notification` as well.

I have to implement new testing schema `for_children_iterations_module`
that should be only used for testing such traversal functions.
I covered most of the cases that `SchemaNode::child()` can return.
It might be overwhelming, but I decided for such coverage to be sure
it is 100% correct as this is critical part that must be correct for
3rd party projects that use `libyang-cpp`-

`SchemaNode::childInstantiables` works as expected, but as the part of
the preparation I also improved tests for this function to
confirm correctness of the traversal algorithm for instantiables children
and to prepare test coverage that can be copy-pasted for testing
`SchemaNode::immediateChildren` and `SchemaNode::childrenDfs`
just with changing expected output.

Change-Id: Ie4cd7c25493d8571378e2b69163fc3c22b59f0e7