API/ABI change: opaque node naming
Our C++ API would ignore the "module name or XML prefix", which turns
out to be *the* relevant part when it comes to opaque node naming. The
prefix is, instead, just that string that might have been inherited from
the parent node when parsing the serialized data; it's an optional
thingy which, if not set explicitly, is implicitly inherited.

Adapt the API for this, and since this *will* break the build, let's
bump the package version.

Change-Id: I199afe5fa7a571034b744531c63b93b9c656563a