build: try building with C++23, if available
We've been using the `CMAKE_CXX_STANDARD_REQUIRED ON` from the very
beginning, at least from cla-sysrepo's commit c2e71252 back in 2016. I
guess that the thinking that I followed was to make sure that if the
compiler is too old, we will get a clean error at CMake time rather than
a build error later on.

However, this actually prevents us from "opportunistically" enabling
support for C++23 *if available*. The way how CMake behaves [1] is
to actually silently downgrade to older versions of the standard. And
this is what we want today: use C++23 if available, and if not, use
C++20.

The only downside is that we'll get no "early error" when building on,
say, a C++14 compiler, but I think I can live with that.

[1] https://cmake.org/cmake/help/latest/prop_tgt/CXX_STANDARD.html

Change-Id: Idc60365f9c1763ae12e2a871a4702e46237c156d