context parseOp add operation types
Added YANG RPC and YANG RPC reply operation types.
Use this when you have a path to the RPC and you have an RPC
input/output in the YANG RPC format - either in JSON or XML (RFC 7950).
Parse the path into a node and then use parseOp to parse the data into
the same node.

Change-Id: If32d78fb79e565739ab6b6b030dd619680a55cfd