build: give up on embedding the git commit identification
We've "always" wanted to embed the tag/commit info into the build, if
only for proper troubleshooting. However, we never documented this, and
someone hit that during their build (see the linked bug). Furthermore,
this tag/commit info was not being used at all in the C++ code, and
there also was no #define in any header. The version info which is in
the pkg-config file comes from a hard-coded version (which is a good
thing to do).

The only use of that version-specific variable was in the Doxygen file,
but there it was not working. Our CMake code tried to be smart and not
trigger a hard rebuild *all the time( (which is a good thing), but that
did not work at all. In CMake, the usual way of achieving this is to
generate a temporary file, and only copy that
always-generated-at-each-build temporary file to a file that's actually
included in some C++ TU when these files are different. However, this
also means that the code which was determining the tag/commit info was
running in a separate CMake process, and as such its variables were
never included in *our* CMake context, and therefore the XXX_VERSION
variable was empty at the time main CMake run, and therefore the
generated Doxyfile had no version info.

Let's just remove that.

I *tried* to make this work via `CMAKE_CONFIGURE_DEPENDS`, but it's a
bit more complex, and one would probably have to use some extra custom
target for generating the Doxyfile "properly". I'm not in the mood of
doing allthat just for a version marker in the docs, but we should
probably fix that later "properly".

Bug: https://github.com/CESNET/libyang-cpp/issues/31
Change-Id: I0c3ca600e2a57e4a54e98262154c75c7b54f6961